merge: incorporate ml/jxl (contract/export/gate) into jxl
Union merge of ml/jxl (ml workspace copy, source last commit:
c367892 fix(jxl): ruff T201 gate + library prints -> logger + meta FALLBACK marker) into ~/cc/py/jxl skeleton (= ~/py/jxl HEAD 9a3cb93).

Composition:
- ml-only 43 files added (contract/, export_yolo_with_contract, d2d_peoplenet,
  ruff.toml, examples/oai/, tests/test_contract, etc.).
- py-only 66 files preserved (det_mine, experiments/, SAMPLES.md, crop_foreground,
  dedup_sem, embed_dino, coreset_image, dedup_image_fp, etc.).
- 101 differ files default to ml version (B7 gate passed).

Manual reconciliation (pyproject.toml):
- Add py-only deps for det_mine validator + rmb synthetic pipeline:
  httpx, albumentations<2.0, imageio[ffmpeg], pillow, transformers, rfdetr.
- Keep ml onnxruntime-gpu, pin <1.28 (1.28+ PyPI has no Linux wheel;
  ml working baseline 1.27.0).
- Keep ml [tool.mypy] strict=false baseline (py was strict=true; tightening
  is a separate phase 2 effort).
- Add py-only [tool.pytest.ini_options] to skip classifier_tch_test
  (imports uninstalled torchsummary, NON-LOCKED-LIB, no replacement path).

Differ triage (101 total): 1 merged + 98 keep-ml + 2 manual (pyproject/README).
- src/jxl/bin/yolo_train.py merged: py adds 6 fine-tuning typer options
  (lr0/lrf/freeze/optimizer/cos_lr/close_mosaic) + project_abs path
  resolution; py is functional superset, ml shebang kept.
- 18 TINY differs (<=4 lines): ml cleaner (portable shebang, frozen dataclass,
  explicit imports vs wildcard, dead-code removed).
- 80 REAL differs: all def/class names match; differences are style/refactor/
  B7-gate fixes (img-train No-Silent-Degradation, public API imports).
- py-only def names (resize_labels/nozero_ratio/test_extract_dir) are dead
  code in py; ml already removed them.

README.md: TODO marker inserted at top (content rewrite is Task 7).

Verification: uv sync OK; import jxl + jxl.contract.embed_contract OK;
jxl_label --help OK; jxl_export_yolo_with_contract --help OK.

Known debt logged in _relocate/LOG.md: 5 person_*/samples_* alias pairs
(ml old-name impl vs py renamed impl) coexist; cleanup deferred to phase 2.

+++ src/jxl/iqa/a.py
@@ -14,7 +14,7 @@ def estimate_clearness(image: ImageNda, ratio: float) -> int:
     # print('hist:', hist.shape)
     hist = hist.flatten()
     # print('hist:', hist.shape)
-    thr = float(ratio * np.sum(hist))
+    thr = ratio * np.sum(hist)
     s = 0.0
     for i, v in enumerate(hist):
         s += float(v)
